requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
openpyxl==3.1.5
python-multipart==0.0.9
python-dotenv==1.0.1
aiosmtplib==3.0.2
requests==2.31.0
